envalysis: Miscellaneous Functions for Environmental Analyses

Small toolbox for data analyses in environmental chemistry and ecotoxicology. Provides, for example, calibration() to calculate calibration curves and corresponding limits of detection (LODs) and limits of quantification (LOQs) according to German DIN 32645 (2008). texture() makes it easy to estimate soil particle size distributions from hydrometer measurements (ASTM D422-63, 2007).

References

Almeida, A. M. D., Castel-Branco, M. M., & Falcao, A. C. (2002). Linear regression for calibration lines revisited: weighting schemes for bioanalytical methods. Journal of Chromatography B, 774(2), 215-222. doi:10.1016/S1570-0232(02)00244-1 .

ASTM D422-63 (2007). Standard Test Method for Particle-Size Analysis of Soils. Technical standard. ASTM International, West Conshohocken, PA. Available from https://www.astm.org/standards/d422.

Borcard D, Gillet F, & Legendre P. (2011). Numerical ecology with R. Springer, New York.

See also

Author

Maintainer: Zacharias Steinmetz z.steinmetz@rptu.de (ORCID)

Other contributors:

Julius Albert (ORCID) [contributor]

Kilian Kenngott (ORCID) [contributor]