ICP-AES calibration data

Sample data of the calibration of silver ions in four-fold replication using inductively coupled plasma atomic emission spectroscopy (ICP-AES).

Format

A data frame containing 16 rows and 6 columns with information on:

Label

sample name

Element

element name

Conc

nominal concentration

Units

concentration units

Signal

signal intensity [cps]

Replicate

number of the respective replicate

See also

Author

Zacharias Steinmetz