Assess matrix effects and matrix-matched calibrations

Calculate the matrix effect by comparing the slope of a solvent-based calibration curve with one or more matrix-matched calibration. The matrix effect is expressed as signal suppression/enhancement ratio.

Usage

matrix_effect(object, ...)

# S3 method for class 'calibration'
matrix_effect(object, ...)

Arguments

object

an object of class 'calibration' obtained from analyzing standard solutions of different concentration (solvent calibration data).

...

additional objects of the same type obtained from matrix-matched calibration data.

Value

The magnitude of a matrix effect is estimated by subtracting the slope of a matrix-matched calibration from that of the solvent-based calibration. The difference is divided by the slope of the solvent-based calibration.

Details

Matrix effects or signal suppression/enhancement ratios should be evaluated during analytical method development to avoid over- or underestimation of sample concentrations. In addition, signal suppression/enhancement ratios may help to justify the validity of a regular solvent calibration as opposed to matrix-matched calibrations. This may be the case if matrix effects or signal suppression/enhancement ratios are close to measurement repeatability.

See also

Author

Julius Albert, Zacharias Steinmetz

Examples

data(din32645)
din <- calibration(Area ~ Conc, data = din32645)

m32645 <- din32645
m32645$Area <- din32645$Area * 1.5
matrix <- calibration(Area ~ Conc, data = m32645)

matrix_effect(din, matrix)
#> din - matrix 
#>          0.5