Degradation of phenolic compounds by Steinmetz et al. (2019)

The sample data is stored in a list consisting of two data.frames: a sequence table (seq) and a sample table (samples).

The sequence table contains gas-chromatography/mass spectrometry measurement data of two phenolic compounds, these are tyrosol and vanillin. Besides the samples, standard mixtures and extraction blanks were acquired in three separate analysis batches. Each measurement resulted in an integrated peak area.

The sample table describes the samples' origin from a 29-day degradation experiment, in which the phenolic compounds were either degraded in the dark by the native soil microbial community or photooxidized under UV irradiation. The samples were processed in threefold replication. Their weight [g], the volume [mL] of extract solution, and the dilution factor were recorded.

Format

A list containing two data.frames.

seq is a data frame with 160 rows and 6 columns with information on:

Compound

name of the phenolic compound

Type

sample type

Batch

number of the extraction batch

Name

sample name

Area

integrated peak area

Spec Conc

specified/nominal concentration [mg/L] of standards

samples is a data frame with 42 rows and 9 columns with information on:

Name

sample name; same as in seq

Day

day of the incubation experiment

Lighting

lighting conditions of the sample (dark/UV)

Sterilization

if the sample was sterilized prior incubation

Treatment

treatment name (Biogradation/Photooxidation)

Replicate

replicate number

Weight

sample weight [g]

Extract

extract volume [mL]

Dilution

dilution factor

References

Steinmetz, Z., Kurtz, M.P., Zubrod, J.P., Meyer, A.H., Elsner, M., & Schaumann, G.E. (2019) Biodegradation and photooxidation of phenolic compounds in soil—A compound-specific stable isotope approach. Chemosphere 230, 210-218. DOI: doi:10.1016/j.chemosphere.2019.05.030 .

See also

Author

Zacharias Steinmetz