Standard error and confidence interval

These wrapper functions compute the standard error (SE) or the confidence interval (CI) of the values in x. If na.rm is TRUE, missing values are removed before the computation proceeds.

Usage

se(x, na.rm = FALSE)

CI(x, level = 0.95, na.rm = FALSE)

Arguments

x

a numeric vector or an R object which is coercible to one by as.vector(x, "numeric").

na.rm

logical. Should missing values be removed?

level

the confidence level required.

See also

Author

Zacharias Steinmetz

Examples

se(1:5)
#> [1] 0.7071068
CI(1:5)
#> [1] 1.385904