Sorption isotherms

This function returns the concentration of a substance sorbed to a surface boundary after an equilibrium has established at constant temperature given the concentration(s) x of the dissolved substance.

Usage

sorption(x, par, type = "freundlich")

Arguments

x

a numeric vector containing the concentration(s) of the dissolved substance.

par

a numeric vector specifying the function parameters, see examples for details and correct order.

type

a character string indicating the type of sorption isotherm to be used: "linear" for the linear type, "freundlich" for the Freundlich isotherm, "langmuir" for the Langmuir isotherm, "BET" for the BET model according to Brunauer, Emmet, and Teller "redlich" for the Redlich-Peterson isotherm.

References

Atkins, P.W. (2001). Physical chemistry, Oxford University Press, Oxford.

Author

Zacharias Steinmetz

Examples

sorption(1:5, par = c(Kd = 2.5), type = "linear")
#> [1]  2.5  5.0  7.5 10.0 12.5
sorption(1:5, par = c(K = 4, n = 0.6), type = "freundlich")
#> [1]  4.000000  6.062866  7.732728  9.189587 10.506111
sorption(1:5, par = c(KL = 2, qmax = 10), type = "langmuir")
#> [1] 6.666667 8.000000 8.571429 8.888889 9.090909
sorption(1:5, par = c(K = 50, qmax = 10, Csat = 10), type = "BET")
#> [1]  9.416196 11.574074 13.648772 16.181230 19.607843
sorption(1:5, par = c(A = 30, B = 0.8), type = "redlich")
#> [1] 15.00000 21.88901 26.40671 29.76609 32.44016