Degradation of phenolic compounds by Steinmetz et al. (2019)

The sample data is stored in a list. It consists of two data.frames: a sequence table (seq) and a sample table (samples).

The sequence table contains gas-chromatography/mass spectrometry measurement data of two phenolic compounds, these are tyrosol and vanillin. Besides the samples, standard mixtures and extraction blanks were acquired in three separate analysis batches. Each measurement resulted in an integrated peak area.

The sample table describes the samples' origin from a 29-day degradation experiment, in which the phenolic compounds were either degraded in the dark by the native soil microbial community or photooxidized under UV irradiation. The samples were processed in threefold replication. Their weight \[g\], the volume \[mL\] of extract solution, and the dilution factor were recorded.

Format

A list containing two data.frames.

References

Steinmetz, Z., Kurtz, M.P., Zubrod, J.P., Meyer, A.H., Elsner, M., Schaumann, G.E. (2019) Biodegradation and photooxidation of phenolic compounds in soil—A compound-specific stable isotope approach. Chemosphere 230, 210-218. DOI: doi:10.1016/j.chemosphere.2019.05.030 .

See also

Author

Zacharias Steinmetz